fix: resolve mypy errors from the sub-1 scale factor and local mode work
CI's Lint & Type Check job runs mypy, which the local dev loop doesn't --
CLAUDE.md's own test instructions only mention pytest and ruff. 17 errors
had accumulated across the local mode and sub-1 scale factor commits,
confirmed absent on main (mypy is clean there), so all are new.

Two real classes of gap, not one:

1. scale.py's compute_guidance()/full_compute_guidance() do scale // N
   arithmetic (scale is float, to allow sub-1 local-mode values like 0.1)
   and pass the float result into ComputeGuidance/DatagenGuidance/
   TrinoGuidance fields typed int. Wrapped each in int() -- these are
   executor/parallelism/replica counts, which are always whole numbers
   by construction regardless of the input scale.

2. Several types were widened for fractional scale in one place but left
   at their old int annotation downstream: DatagenConfig.get_effective_
   scale() -> int (schema.py) feeds compute_guidance(scale: float)
   directly; WizardState.scale, BenchmarkResult.scale, and
   BenchmarkMetrics.scale were all still int while every value flowing
   into them already carries the real fractional scale. Widened all four
   to float rather than truncating -- a local-mode run at scale 0.1
   should report 0.1 in its results, not silently round to 0.

Two smaller fixes surfaced by the same trace:
- init_wizard.py's config-YAML regex substitution (scale:\s*\d+) would
  silently fail to match a fractional scale, leaving the template's
  default in place with no error. Reachable via `lakebench init --scale
  0.1` on the non-local wizard path (CLI override applied after the
  wizard's own int-only prompt returns). Widened the pattern to match a
  decimal point.
- cli/__init__.py's status command checked `if config_file:` to decide
  whether cfg (separately assigned, starts as None) was safe to use --
  correct in practice since the load path re-raises on failure, but
  mypy can't see the connection between the two variables and neither
  can a reader tracing it cold. Checks `cfg is not None` directly now.

_local.py's status dict was typed dict[str, object], which is correct
for values never read back, but this dict's own consumers call .get()
and iterate its values -- dict[str, Any] describes what the code
actually depends on.

All fixes verified against mypy 1.19.1 (matching CI's resolution) and the
full suite: 1576 passed, ruff clean.

+++ src/lakebench/cli/_run.py
@@ -929,7 +929,9 @@ def run(
                         _datagen_output_gb = dg_info.size_bytes / (1024**3)
                     if dg_info.object_count:
                         # Estimate rows from scale factor (1.5M rows per scale unit)
-                        _datagen_output_rows = cfg.architecture.workload.datagen.scale * 1_500_000
+                        _datagen_output_rows = int(
+                            cfg.architecture.workload.datagen.scale * 1_500_000
+                        )
                 except Exception as e:
                     logger.warning("Could not measure bronze bucket size: %s", e)
             except Exception as e:
